+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## 0.3.1 - 2019-03-25
 ### Changed
 - refactoring: getQuota without webdav client
 - refactoring: createFolder without webdav client